Move common UI code to common.ui
