+++ wearos/src/main/kotlin/com/boswelja/smartwatchextensions/batterysync/PhoneBatteryComplicationProvider.kt
@@ -17,7 +17,7 @@ import androidx.wear.complications.datasource.ComplicationRequest
 import com.boswelja.smartwatchextensions.ActionsActivity
 import com.boswelja.smartwatchextensions.R
 import com.boswelja.smartwatchextensions.common.batterysync.References.REQUEST_BATTERY_UPDATE_PATH
-import com.boswelja.smartwatchextensions.common.getBatteryResource
+import com.boswelja.smartwatchextensions.common.ui.getBatteryResource
 import com.boswelja.smartwatchextensions.phoneStateStore
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
