+++ wearos/src/main/kotlin/com/boswelja/smartwatchextensions/batterysync/ui/BatteryStatsCard.kt
@@ -10,8 +10,8 @@ import androidx.wear.compose.material.Icon
 import androidx.wear.compose.material.MaterialTheme
 import androidx.wear.compose.material.Text
 import com.boswelja.smartwatchextensions.R
-import com.boswelja.smartwatchextensions.common.BatteryIcon
-import com.boswelja.smartwatchextensions.common.ExtensionCard
+import com.boswelja.smartwatchextensions.common.ui.BatteryIcon
+import com.boswelja.smartwatchextensions.common.ui.ExtensionCard
 
 @Composable
 fun BatteryStatsCard(
